refactor(telemetry): file the generic bus, timer and logger under utils
managers/ is for the stateful machinery a client owns. Three of its seven
files were not that: scheduler is a debounced timer with no telemetry
knowledge, logger holds no state, and hooks is a generic event bus whose
own types already lived in utils/hooks.ts — so the bus joins them there
rather than sitting one folder away from its contract.

userInfo stays a manager: it holds the user id and identify() is callable
before start(), so the plugin shape it resembles would drop early calls.
